requirements.txt
gitpython==0.3.2.RC1
radon==0.5.3
pylint==1.3.0
South==1.0
django-social-auth==0.7.28
celery==3.1.13
django-celery==3.1.10
